Combine similar API route classes
