+++ src/php/api/routes/elections/class.elections.by.id.php
@@ -1,6 +1,16 @@
 <?php
 
 namespace API {
+    
+    class ElectionsByBioguideId extends ElectionsRoute {
+        public static function parameters() { return ["id"]; }
+        
+        public static function fetchResult() {
+            $id = Parameters::get("id");
+            return \AuditCongress\MemberElections::getByBioguideId($id);
+        }
+    }
+ 
     class ElectionsByFecId extends ElectionsRoute {
 
         public static function parameters() { return ["fecid"]; }
