+++ src/php/api/routes/offices/class.offices.by.id.php
@@ -1,6 +1,16 @@
 <?php
 
 namespace API {
+    class OfficesByOfficeId extends OfficesRoute {
+
+        public static function parameters() { return ["id"]; }
+        
+        public static function fetchResult() {
+            $id = Parameters::get("id");
+            return \AuditCongress\MemberOffices::getById($id);
+        }
+    }
+
     class OfficesByBioguideId extends OfficesRoute {
 
         public static function parameters() { return ["bioguideId"]; }
