--- src/php/api/routes/offices/class.offices.by.office.id.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace API {
-    class OfficesByOfficeId extends OfficesRoute {
-
-        public static function parameters() { return ["id"]; }
-        
-        public static function fetchResult() {
-            $id = Parameters::get("id");
-            return \AuditCongress\MemberOffices::getById($id);
-        }
-    }
-}
-
-?>
